requirements.txt
# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
requests==2.31.0
fake-useragent==1.3.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
sqlalchemy==2.0.23
pandas==2.1.3

# AI/ML
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
spacy==3.7.2
statsmodels==0.14.0
prophet==1.1.4
tensorflow==2.15.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Utilities
tqdm==4.66.1
python-multipart==0.0.6
aiofiles==23.2.1